Implemento la llamda a la API en la actividad que muestra las varitas. Por hacer la creación/rotura de varitas

+++ Trimestre-02/MundoMagico/app/src/main/java/com/alejandro/mundomagico/model/CreateVaritaDTO.kt
@@ -2,5 +2,5 @@ package com.alejandro.mundomagico.model
 
 import java.math.BigDecimal
 
-class CreateVaritaDTO(val madera: String, val nucleo: String, val longitud: BigDecimal) {
+class CreateVaritaDTO(val madera: String, val nucleo: String, val longitud: Float) {
 }
